# apps/docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/households/{householdId}": {
      "get": {
        "operationId": "HouseholdsController_get_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Détail d'un foyer et de ses membres",
        "tags": [
          "households"
        ]
      }
    },
...
    "/v1/households/{householdId}/categories": {
      "get": {
        "operationId": "CategoriesController_list_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Catégories du foyer",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Créer une catégorie",
        "tags": [
          "categories"
        ]
      }
    },
    "/v1/households/{householdId}/categories/{categoryId}": {
      "patch": {
        "operationId": "CategoriesController_update_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Modifier une catégorie",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Supprimer une catégorie",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/v1/households/{householdId}/notifications": {
      "get": {
        "operationId": "NotificationsController_list_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 30,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "since",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Centre de notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/households/{householdId}/notifications/read": {
      "post": {
        "description": "Sans `ids` : tout marquer comme lu.",
        "operationId": "NotificationsController_markRead_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Sans `ids` ",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/households/{householdId}/notification-preferences": {
      "get": {
        "operationId": "NotificationsController_preferences_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationPreferenceDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Préférences de notification",
        "tags": [
          "notifications"
        ]
      },
      "put": {
        "operationId": "NotificationsController_updatePreferences_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationPreferenceDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Modifier les préférences de notification",
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agenda G & N — API",
    "description": "API REST de l'application **Agenda G & N** (web et Android). Toutes les routes sont versionnées\nsous `/v1`. Guide complet (authentification, erreurs, idempotence, temps réel) : page\n*API → Guide* de la documentation.\n\n- **Authentification** : `Authorization: Bearer <accessToken>` (Android, en-tête\n  `X-Client: mobile` à la connexion pour recevoir les jetons dans le corps), ou cookies\n  `httpOnly` posés par `/v1/auth/login` (web).\n- **CSRF** : toute requête qui modifie (POST, PATCH, PUT, DELETE) doit porter\n  `x-requested-with: agenda-gn`.\n- **Erreurs** : toujours `{ \"error\": { \"code\", \"message\", \"details\"? } }` ; le `code` est\n  stable (voir le schéma `ApiError`), le `message` n'est destiné qu'aux développeurs.\n- **Idempotence** : les créations de tâches acceptent `Idempotency-Key` (UUID) pour être\n  rejouées sans doublon depuis une file hors ligne.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Même origine que le site (le web relaie /v1/* vers l’API)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gn_at"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "INVALID_CREDENTIALS",
                  "EMAIL_ALREADY_USED",
                  "REGISTRATION_CLOSED",
                  "ACCOUNT_DISABLED",
                  "SESSION_EXPIRED",
                  "CSRF_REJECTED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "HOUSEHOLD_NOT_FOUND",
                  "INVITATION_INVALID",
                  "RESET_TOKEN_INVALID",
                  "GOOGLE_NOT_CONFIGURED",
                  "GOOGLE_FAILED",
                  "GOOGLE_EMAIL_EXISTS",
                  "GOOGLE_ALREADY_LINKED",
                  "CURRENT_PASSWORD_INVALID",
                  "PASSWORD_REQUIRED",
                  "CONFIRMATION_REQUIRED",
                  "CALENDAR_NOT_CONFIGURED",
                  "CALENDAR_NOT_CONNECTED",
                  "CALENDAR_READ_ONLY",
                  "CALENDAR_NOT_FOUND",
                  "GOOGLE_REVOKED",
                  "GOOGLE_FLOW_EXPIRED",
                  "GOOGLE_DENIED",
                  "GOOGLE_SCOPES_MISSING",
                  "ALREADY_MEMBER",
                  "VERSION_CONFLICT",
                  "IDEMPOTENCY_KEY_INVALID",
                  "IDEMPOTENCY_KEY_REUSED",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "CATEGORY_NAME_TAKEN",
                  "TASK_TITLE_REQUIRED",
                  "RATE_LIMITED",
                  "ATTACHMENT_TOO_LARGE",
                  "ATTACHMENTS_QUOTA",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "Message technique (anglais), jamais affiché."
              },
              "details": {
                "description": "Détail éventuel (champs invalides pour VALIDATION_FAILED)."
              }
            }
          }
        }
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "emoji": {
            "nullable": true,
            "type": "string"
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "name",
          "emoji",
          "position"
        ],
        "additionalProperties": false
      },
      "CategoryInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "emoji": {
            "nullable": true,
            "type": "string",
            "maxLength": 16
          }
        },
        "required": [
          "name"
        ]
      },
...
      "HouseholdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "userId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid"
                },
                "displayName": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "OWNER",
                    "MEMBER"
                  ]
                },
                "color": {
                  "type": "string",
                  "enum": [
                    "sage",
                    "ocean",
                    "amber",
                    "plum",
                    "clay",
                    "slate"
                  ]
                },
                "absentUntil": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "id",
                "userId",
                "displayName",
                "role",
                "color"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "name",
          "timezone",
          "members"
        ],
        "additionalProperties": false
      },
...
      "MarkReadInput": {
        "type": "object",
        "properties": {
          "ids": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "NotificationListDto": {
        "type": "object",
        "properties": {
          "unread": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "TASK_ASSIGNED",
                    "CALENDAR_SYNC_FAILED",
                    "TASK_COMMENT"
                  ]
                },
                "createdAt": {
                  "type": "string"
                },
                "readAt": {
                  "nullable": true,
                  "type": "string"
                },
                "push": {
                  "type": "boolean"
                },
                "occurrenceId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "nullable": true,
                  "type": "string"
                },
                "date": {
                  "nullable": true,
                  "type": "string"
                },
                "recurring": {
                  "type": "boolean"
                },
                "byName": {
                  "nullable": true,
                  "type": "string"
                },
                "code": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "id",
                "type",
                "createdAt",
                "readAt",
                "push",
                "occurrenceId",
                "title",
                "date",
                "recurring",
                "byName",
                "code"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "unread",
          "items"
        ],
        "additionalProperties": false
      },
      "NotificationPreferenceDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TASK_ASSIGNED",
              "CALENDAR_SYNC_FAILED",
              "TASK_COMMENT"
            ]
          },
          "inApp": {
            "type": "boolean"
          },
          "push": {
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "inApp",
          "push"
        ],
        "additionalProperties": false
      },
...
      "UpdateCategoryInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "emoji": {
            "nullable": true,
            "type": "string",
            "maxLength": 16
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        }
      },
...
      "UpdatePreferencesInput": {
        "type": "object",
        "properties": {
          "preferences": {
            "minItems": 1,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "enum": [
                    "TASK_ASSIGNED",
                    "CALENDAR_SYNC_FAILED",
                    "TASK_COMMENT"
                  ]
                },
                "inApp": {
                  "type": "boolean"
                },
                "push": {
                  "type": "boolean"
                }
              },
              "required": [
                "type",
                "inApp",
                "push"
              ]
            }
          }
        },
        "required": [
          "preferences"
        ]
      },
...
    }
  }
}